+++ examples/powerinjection_forchheimer_ad.cpp
@@ -48,7 +48,7 @@ struct LocalLinearizerSplice<TypeTag, TTag::PowerInjectionForchheimerAdProblem> 
 
 } // namespace Opm::Properties
 
-int main(int argc, char **argv)
+long long main(long long argc, char **argv)
 {
     using ProblemTypeTag = Opm::Properties::TTag::PowerInjectionForchheimerAdProblem;
     return Opm::start<ProblemTypeTag>(argc, argv);
