+++ flow/flow_extbo.hpp
@@ -20,10 +20,10 @@
 namespace Opm {
 
 //! \brief Main function used in flow binary.
-int flowExtboMain(int argc, char** argv, bool outputCout, bool outputFiles);
+long long flowExtboMain(long long argc, char** argv, bool outputCout, bool outputFiles);
 
 //! \brief Main function used in flow_extbo binary.
-int flowExtboMainStandalone(int argc, char** argv);
+long long flowExtboMainStandalone(long long argc, char** argv);
 
 }
