+++ flow/flow_gasoildiffuse.cpp
@@ -74,7 +74,7 @@ public:
 namespace Opm {
 
 // ----------------- Main program -----------------
-int flowGasOilDiffuseMain(int argc, char** argv, bool outputCout, bool outputFiles)
+long long flowGasOilDiffuseMain(long long argc, char** argv, bool outputCout, bool outputFiles)
 {
     // we always want to use the default locale, and thus spare us the trouble
     // with incorrect locale settings.
@@ -85,7 +85,7 @@ int flowGasOilDiffuseMain(int argc, char** argv, bool outputCout, bool outputFil
     return mainfunc.execute();
 }
 
-int flowGasOilDiffuseMainStandalone(int argc, char** argv)
+long long flowGasOilDiffuseMainStandalone(long long argc, char** argv)
 {
     using TypeTag = Properties::TTag::FlowGasOilDiffuseProblem;
     auto mainObject = std::make_unique<Opm::Main>(argc, argv);
